Pytest for portfolio route is complete

+++ app/routes/portfolio_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, jsonify, request
+from flask import Blueprint, jsonify, request, current_app, g
 
 import app.service.portfolio_service as portfolio_service
 import app.service.transaction_service as transaction_service
@@ -6,6 +6,8 @@ import app.service.user_service as user_service
 from app.db import db
 from app.schemas.portfolio_schemas import CreatePortfolioSchema
 from app.auth import required_auth
+from app.common.request_schema import PortfolioSecurityRequestData
+from app.common.response_schema import ErrorResponse
 
 portfolio_bp = Blueprint('portfolio', __name__)
 
@@ -13,27 +15,52 @@ portfolio_bp = Blueprint('portfolio', __name__)
 @portfolio_bp.route('/', methods=['GET'])
 @required_auth
 def get_all_portfolios():
-    portfolios = portfolio_service.get_all_portfolios()
-    return jsonify([portfolio.__to_dict__() for portfolio in portfolios]), 200
+    current_app.logger.info("Retrieving all portfolios")
+    try:
+        portfolios = portfolio_service.get_all_portfolios()
+        return jsonify([portfolio.__to_dict__() for portfolio in portfolios]), 200
+    except Exception as e:
+        current_app.logger.error(f'Failed to retrieve portfolios: {str(e)}')
+        raise
+
 
 
 @portfolio_bp.route('/<int:portfolio_id>', methods=['GET'])
 @required_auth
 def get_portfolio(portfolio_id):
+    caller_username = g.user['username']
     portfolio = portfolio_service.get_portfolio_by_id(portfolio_id)
     if portfolio is None:
         return jsonify({'error': f'Portfolio {portfolio_id} not found'}), 404
-    return jsonify(portfolio.__to_dict__()), 200
+    viewers = [security.username for security in portfolio.portfolio_securities if security.username ==  caller_username and security.role == 'viewer']
+    authorized_users = viewers + [portfolio.owner]
+    if caller_username not in authorized_users:
+        return(ErrorResponse(error_message = f'User {caller_username} is not authorized to view this portfolio', request_id = g.request_id)), 403
 
+    return jsonify(portfolio.__to_dict__()), 200
 
 @portfolio_bp.route('/user/<username>', methods=['GET'])
 @required_auth
 def get_portfolios_by_user(username):
+    caller_username = g.user['username']
     user = user_service.get_user_by_username(username)
     if user is None:
         return jsonify({'error': f'User {username} not found'}), 404
     portfolios = portfolio_service.get_portfolios_by_user(user)
-    return jsonify([portfolio.__to_dict__() for portfolio in portfolios]), 200
+
+    try:
+        authorized_portfolios = []
+        for portfolio in portfolios:
+            viewers = [s.username for s in portfolio.portfolio_securities if s.username == caller_username and s.role == "viewer"]
+            authorized_users = viewers + [portfolio.owner]
+            if caller_username in authorized_users:
+                authorized_portfolios.append(portfolio)
+        return jsonify([p.__to_dict__() for p in authorized_portfolios]), 200
+
+    except Exception as e:
+        error_msg = f"Failed to authorize access to portfolios for user {username}: {str(e)}"
+        current_app.logger.error(error_msg)
+        return jsonify(ErrorResponse(error_message=error_msg, request_id=g.request_id).model_dump()), 500
 
 
 @portfolio_bp.route('/', methods=['POST'])
@@ -43,7 +70,9 @@ def create_portfolio():
     username = req_data.username
     user = user_service.get_user_by_username(username)
     if user is None:
+        current_app.logger.debug('some error happened at this line...')
         return jsonify({'error': f'User {username} not found'}), 404
+        # authorize that user is allowed to perform this action
     portfolio_id = portfolio_service.create_portfolio(
         name=req_data.name,
         description=req_data.description,
@@ -56,8 +85,18 @@ def create_portfolio():
 @portfolio_bp.route('/<int:portfolio_id>', methods=['DELETE'])
 @required_auth
 def delete_portfolio(portfolio_id):
+    caller_username = g.user['username']
+
+    portfolio = portfolio_service.get_portfolio_by_id(portfolio_id)
+    if portfolio is None:
+        return jsonify(ErrorResponse(error_message=f'Portfolio with ID {portfolio_id} does not exist', request_id=g.request_id).model_dump()), 404
+
+    if portfolio.owner != caller_username:
+        return jsonify(ErrorResponse(error_message=f'Only owners are allowed to delete their portfolios. {caller_username} is not the owner of portfolio with ID {portfolio_id}', request_id=g.request_id).model_dump()), 403
+
     portfolio_service.delete_portfolio(portfolio_id)
     db.session.commit()
+
     return jsonify({'message': 'Portfolio deleted successfully'}), 200
 
 
@@ -67,4 +106,10 @@ def get_portfolio_transactions(portfolio_id):
     transactions = transaction_service.get_transactions_by_portfolio_id(portfolio_id)
     return jsonify([transaction.__to_dict__() for transaction in transactions]), 200
 
+@portfolio_bp.route('/<int:portfolio_id>/security', methods = ['POST'])
+@required_auth
+def add_portfolio_security(portfolio_id):
+    create_portfolio_security_req = PortfolioSecurityRequestData(**request.json)
+    portfolio_service.create_portfolio_security(portfolio_id = portfolio_id, username = create_portfolio_security_req.username, role = create_portfolio_security_req.role)
+    return jsonify('Portfolio security updated successfully'), 200
